优化

+++ TFY_TabBarController/TFY_Class/TestVC/TestGetTabBarItemVC.m
@@ -30,7 +30,7 @@
 //    TfySY_TabBarItem *item = tabBarVC.tfySY_TabBar.tabBarItems[2]; // 这个下标是固定的，代表这个页面下的item
     
     self.badge ++;
-    item.badge = [NSString stringWithFormat:@"%ld",self.badge];
+    item.badge = [NSString stringWithFormat:@"%ld",(long)self.badge];
     // 为0是否自动隐藏
 //    item.badgeLabel.automaticHidden = YES;
